# site/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Toren intake API",
    "version": "0.1.1",
    "description": "The HTTP API every Toren deployment serves (toren dev locally, the load balancer on AWS). Toren is self-hosted: this API lives at YOUR deployment's URL, not at toren.run. Trigger durable agent runs, read status and full event transcripts, resolve approvals, upload files, and hold sessions. All endpoints except /healthz require a bearer token: the deployment's admin token or an issued API key (issued keys cannot manage keys, schedules, or invites).",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:7433",
      "description": "toren dev default; replace with your deployment's URL"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "TOREN_API_TOKEN (admin) or an issued API key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SessionTurn": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ]
          },
          "text": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "seq": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/runs": {
      "post": {
        "operationId": "startRun",
        "summary": "Start a durable run; 202 immediately, poll GET /runs/{runId}",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "input"
                ],
                "properties": {
                  "input": {
                    "type": "string"
                  },
                  "agent": {
                    "type": "string",
                    "description": "defaults to the deployment's default agent"
                  },
                  "process": {
                    "type": "string",
                    "description": "a named process of the agent; defaults to its default_process, else main"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "file ids from POST /files"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "run created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string"
                    },
                    "agent": {
                      "type": "string"
                    },
                    "process": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "unknown agent or process (the error lists what exists), or bad body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listRuns",
        "summary": "Newest 50 runs per agent crew",
        "responses": {
          "200": {
            "description": "runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RunSummary"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/runs/{runId}/cancel": {
      "post": {
        "operationId": "cancelRun",
        "summary": "Retire a run: retries stop and queued work for it becomes a no-op",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cancelled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "unknown or already terminal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{runId}/approvals": {
      "post": {
        "operationId": "resolveApproval",
        "summary": "Approve or deny a parked tool call; the run wakes and continues",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "taskId",
                  "stepId",
                  "granted"
                ],
                "properties": {
                  "taskId": {
                    "type": "string"
                  },
                  "stepId": {
                    "type": "string"
                  },
                  "granted": {
                    "type": "boolean"
                  },
                  "comment": {
                    "type": "string"
                  },
                  "by": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "resolved"
          },
          "404": {
            "description": "unknown run or approval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions": {
      "post": {
        "operationId": "startSession",
        "summary": "Open a durable conversation with an agent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "agent": {
                    "type": "string"
                  },
                  "channel": {
                    "type": "string"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "session started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string"
                    },
                    "agent": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listSessions",
        "summary": "Recent conversations across every channel",
        "responses": {
          "200": {
            "description": "sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RunSummary"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{runId}": {
      "get": {
        "operationId": "getSession",
        "summary": "Transcript and state (working | awaiting_input | completed | failed | cancelled)",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "session detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string"
                    },
                    "agent": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "transcript": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SessionTurn"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{runId}/messages": {
      "post": {
        "operationId": "sendSessionMessage",
        "summary": "Send the next turn; 409 while the agent is mid-turn (strict turn-taking); close: true ends the session",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "channel": {
                    "type": "string"
                  },
                  "close": {
                    "type": "boolean"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "accepted"
          },
          "409": {
            "description": "agent is mid-turn; wait for awaiting_input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedules/{id}": {
      "delete": {
        "operationId": "deleteSchedule",
        "summary": "Delete a schedule (admin token only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deleted"
          },
          "404": {
            "description": "no such schedule"
          }
        }
      }
    },
...
    "/keys/{id}": {
      "delete": {
        "operationId": "revokeApiKey",
        "summary": "Revoke a key immediately (admin token only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "revoked"
          }
        }
      }
    },
...
  }
}